api-case-studies/g4ae/requirements.txt
requests==2.31.0
pandas==2.2.0
python-dotenv==1.0.1
openpyxl==3.1.2
xlsxwriter==3.1.9